in progres edit competitor

+++ belt3000-back/src/competitor/competitor.controller.js
@@ -6,13 +6,47 @@ const withAuth = require('../auth/auth.middleware');
 
 router.get('/all', withAuth, async (req, res) => {
   const competitors = await Competitor.find({});
-  console.log(competitors);
-  return res.json(competitors);
+
+  return res.status(200).json(competitors);
+});
+
+router.get('/one/:id', withAuth, async (req, res) => {
+  if (!req || !req.params || !req.params.id) {
+    return res.status(400).json('Invalid param.');
+  }
+  const competitor = await Competitor.findById(req.params.id);
+
+  return res.status(200).json(competitor);
 });
 
 router.delete('/:id', withAuth, async (req, res) => {
+  if (!req || !req.params || !req.params.id) {
+    return res.status(400).json('Invalid param.');
+  }
+
   await Competitor.findOneAndDelete({ _id: req.params.id });
   return res.status(200).send();
+});
+
+router.patch('/:id', withAuth, validateAddCompetitor, async (req, res, next) => {
+  if (!req || !req.params || !req.params.id) {
+    return res.status(400).json('Invalid param.');
+  }
+
+  const competitor = await Competitor.findById(req.params.id);
+  competitor.firstname = req.body.firstname;
+  competitor.lastname = req.body.lastname;
+  competitor.isAdult = Boolean(req.body.isAdult);
+  competitor.belt = req.body.belt;
+  competitor.stripes = Number(req.body.stripes);
+
+  try {
+    await competitor.save();
+  } catch (e) {
+    return next(e.toString());
+  }
+
+  return res.status(200).json(competitor);
 });
 
 router.post('/add', withAuth, validateAddCompetitor, async (req, res, next) => {
